Move progressbar code to its own module
