+++ .github/workflows/tests.yml
@@ -69,6 +69,7 @@ jobs:
             tests/distributions/test_shape_utils.py
             tests/distributions/test_mixture.py
             tests/test_testing.py
+            tests/test_progress_bar.py
 
           - |
             tests/distributions/test_continuous.py
