+++ pymc/sampling/parallel.py
@@ -33,10 +33,9 @@ from threadpoolctl import threadpool_limits
 from pymc.backends.zarr import ZarrChain
 from pymc.blocking import DictToArrayBijection
 from pymc.exceptions import SamplingError
+from pymc.progress_bar import ProgressBarManager, default_progress_theme
 from pymc.util import (
-    ProgressBarManager,
     RandomGeneratorState,
-    default_progress_theme,
     get_state_from_generator,
     random_generator_from_state,
 )
