+++ tests/test_progress_bar.py
@@ -1,0 +1,46 @@
+#   Copyright 2025 - present The PyMC Developers
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+import pymc as pm
+
+
+def test_progressbar_nested_compound():
+    # Regression test for https://github.com/pymc-devs/pymc/issues/7721
+
+    with pm.Model():
+        a = pm.Poisson("a", mu=10)
+        b = pm.Binomial("b", n=a, p=0.8)
+        c = pm.Poisson("c", mu=11)
+        d = pm.Dirichlet("d", a=[c, b])
+
+        step = pm.CompoundStep(
+            [
+                pm.CompoundStep([pm.Metropolis(a), pm.Metropolis(b), pm.Metropolis(c)]),
+                pm.NUTS([d]),
+            ]
+        )
+
+        kwargs = {
+            "draws": 10,
+            "tune": 10,
+            "chains": 2,
+            "compute_convergence_checks": False,
+            "step": step,
+        }
+
+        # We don't parametrize to avoid recompiling the model functions
+        for cores in (1, 2):
+            pm.sample(**kwargs, cores=cores, progressbar=True)  # default is split+stats
+            pm.sample(**kwargs, cores=cores, progressbar="combined")
+            pm.sample(**kwargs, cores=cores, progressbar="split")
+            pm.sample(**kwargs, cores=cores, progressbar=False)
